refactor(bar2vtk): Move argparse code to module function
 - Moved all the argparse code to a separate function
