+++ vtkpytools/barfiletools/__init__.py
@@ -2,6 +2,6 @@ from .core import form2DGrid, computeEdgeNormals
 from .data import (binaryVelbar, binaryStsbar, calcCf, calcReynoldsStresses,
                    compute_vorticity, sampleDataBlockProfile, calcWallShearGradient,
                    wallAlignRotationTensor)
-from .bar2vtk import bar2vtk
+from .bar2vtk import bar2vtk, bar2vtk_parse
 
 __name__ = "barfiletools"
